[MAIN] Update project files and enhance error handling
- Add .idea/ to .gitignore
- Update Ruby version to 3.3.3
- Fix file deletion logic in city-state.rb
- Bump version to 1.1.2 in version.rb

+++ lib/city-state.rb
@@ -67,7 +67,7 @@ module CS
     @countries = {}
     @states = {}
     @cities = {} # invalidades cache
-    File.delete COUNTRIES_FN # force countries.yml to be generated at next call of CS.countries
+    File.delete COUNTRIES_FN if File.exist? COUNTRIES_FN # force countries.yml to be generated at next call of CS.countries
     true
   end
